still build

+++ smart_contract/src/constants/claim.cairo
@@ -3,7 +3,7 @@ use smart_contract::constants::claim_status::ClaimStatus;
 
 #[derive(Drop, Serde, starknet::Store)]
 pub struct Claim {
-    pub id: u8,
+    pub id: u128,
     pub policy_holder: ContractAddress,
     pub claim_amount: u256,
     pub claim_details: ByteArray,
